requirements_face.txt
insightface==0.7.3
onnxruntime==1.16.0
opencv-python>=4.8.0
numpy<2.0,>=1.24.0
scipy>=1.11.0
flask>=3.0.0
flask-cors>=4.0.0
Pillow>=10.0.0
psycopg2-binary>=2.9.9
python-dotenv>=1.0.0
tqdm>=4.66.1

